requirements.txt
# Веб-фреймворк
Flask>=2.3.0
Werkzeug>=2.3.0

# Работа с Excel и данными
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.0

# Работа с PDF
pdfplumber>=0.9.0

# Визуализация
plotly>=5.15.0
matplotlib>=3.7.0

# Дополнительные зависимости
python-dotenv>=1.0.0

# Яндекс.GPT SDK
yandex-cloud-ml-sdk>=0.1.0

# GigaChat SDK
gigachat>=0.1.0

# PDF генерация
reportlab>=4.0.0
Pillow>=10.0.0


